requirements.txt
Python 3.7+
PyTorch 1.7+
torchvision
numpy
matplotlib